requirements.txt
streamlit>=1.36
pandas>=2.2
plotly>=5.22
python-dateutil>=2.9